refactor: #18 extract MachineContactMaterialRegistrar base for oriented friction
Move the machine-agnostic per-instance oriented-friction registration logic out
of DumpTruckContactMaterialRegistrar into a new abstract base
MachineContactMaterialRegistrar. DumpTruckContactMaterialRegistrar becomes a
thin subclass (script GUID unchanged, so the ic120 prefab wiring is preserved),
and ExcavatorContactMaterialRegistrar is added for the upcoming excavator wiring.

No behavioural change; prepares per-machine reuse (Excavator next).

Refs #18
