+++ Assets/Scripts/ExcavatorContactMaterialRegistrar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8c7faf2cd94e2f644b0443951c992ae8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
