+++ Assets/Scripts/MachineContactMaterialRegistrar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dbb58e5ff7316b045a581a7b742352f3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
